Specified a Generic Relation from BaseModel class to DataCompliance class so that if an object is deleted, its associated data compliance objects will also be deleted and removed DeleteOrphanedDataComplianceData job
