+++ nautobot/extras/datasources/git.py
@@ -1050,6 +1050,14 @@ def delete_git_graphql_queries(repository_record, job_result, preserve=None):
                 job_result.log(error_msg, level_choice=LogLevelChoices.LOG_ERROR, grouping="graphql queries")
 
 
+def refresh_git_data_compliance_rules(repository_record, job_result, delete=False):  # pylint: disable=W0613
+    """Callback for repo refresh."""
+    from nautobot.nautobot_data_validation_engine.custom_validators import get_classes_from_git_repo
+
+    job_result.log("Successfully pulled git repo", level_choice=LogLevelChoices.LOG_INFO)
+    for compliance_class in get_classes_from_git_repo(repository_record):
+        job_result.log(f"Found class {compliance_class.__name__!s}", level_choice=LogLevelChoices.LOG_INFO)
+
 # Register built-in callbacks for data types potentially provided by a GitRepository
 register_datasource_contents(
     [
@@ -1103,5 +1111,14 @@ register_datasource_contents(
                 callback=refresh_git_graphql_queries,
             ),
         ),
+        (
+        "extras.gitrepository",
+        DatasourceContent(
+            name="data compliance rules",
+            content_identifier="nautobot_data_validation_engine.data_compliance_rules",
+            icon="mdi-file-document-outline",
+            callback=refresh_git_data_compliance_rules,
+        ),
+    )
     ]
 )
